Allow `Relation<T>` to be joined with `Variable<(T, S)>`
This case occurred in `Polonius`, where we had to create a separate
`Relation<(T, ())>` to make things work.

Implementing this required a change to the interface of `JoinInput`,
since we are not allowed to assume that a `Relation<T>` and a
`Relation<(T, ())>` are layout compatible. Now, `stable` takes a
callback instead of returning a slice directly.